feat: Add handling for SQS RDS event
