+++ serverless/database-backup/src/event.ts
@@ -1,0 +1,16 @@
+import { RdsEvent } from './interfaces'
+
+export const RDS_EVENTS = {
+  BACKUP_COMPLETE: 'RDS-EVENT-0002',
+}
+
+export const parseRdsEvents = (event: any): Array<RdsEvent> => {
+  return event.Records.map((record: any) => {
+    const snsMessage = JSON.parse(record.Sns.Message)
+    const eventId = snsMessage['Event ID'].split('#').pop()
+    return {
+      eventId,
+      message: snsMessage['Event Message'],
+    }
+  })
+}
